requirements.txt
scikit-learn==1.6.1
pandas==2.2.2
numpy==1.26.4
plotly==5.24.1
umap-learn==0.5.7
tqdm==4.67.1
kaleido